Little update
Change font of the button so it stands out more, the application it self is a little bigger. NEW ICON... Changed the icon so it doesn't look like you got it from Walmart. The font for the little formula is now smaller so its not all in your face, and is italicized to show "instruction."